+++ CHANGELOG.md
@@ -19,7 +19,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-(None)
+### Added
+
+- Added tasks for selecting/deselecting Checkboxes/Radio buttons
 
 
 ## [1.2.2] - 2021-07-15